+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Is ignored Changelog
 
 ## 1.0.0 (2020-xx-xx)
+#### New Feature
+- check which files in directory are excluded or not excluded in ignore files
+- use patterns for `npm` and `git` to check which files should or shouldn't be ignored
+- add or remove files / directories from ignore files